Manpage: Add mention of mouse acceleration with zero threshold
xset m A/B 0 gives a different behaviour, smooth acceleration, than typical
A/B C (C not zero) which is two speed. This behaviour is, for some people,
nice. But the point is that it is not documented in man page, so nobody
can just test it except if discovered due some reason more complex than read
the docs.